backend/sdlcctl/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sdlcctl"
version = "1.7.0"
description = "SDLC 6.0.6 CLI Orchestrator with E2E API Testing and OPA Integration"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "SDLC Orchestrator Team", email = "sdlc@example.com"}
]
maintainers = [
    {name = "SDLC Orchestrator Team", email = "sdlc@example.com"}
]
keywords = ["sdlc", "validation", "cli", "documentation", "structure"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Quality Assurance",
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "click<8.2",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
sdlcctl = "sdlcctl.cli:run"

[project.urls]
Homepage = "https://github.com/sdlc-orchestrator/sdlcctl"
Documentation = "https://docs.sdlc-orchestrator.com/cli"
Repository = "https://github.com/sdlc-orchestrator/sdlcctl"
Issues = "https://github.com/sdlc-orchestrator/sdlcctl/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["sdlcctl*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["sdlcctl"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
fail_under = 95

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C4", "SIM"]
ignore = ["D100", "D104", "D203", "D213"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]
